Agregando Response Entity en los controladores

+++ src/main/java/com/antonativa/antonativa/controllers/ProductoController.java
@@ -4,6 +4,8 @@ import com.antonativa.antonativa.models.Producto;
 import com.antonativa.antonativa.models.Settings;
 import com.antonativa.antonativa.repository.ProductoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDateTime;
@@ -13,20 +15,22 @@ import java.util.Collection;
 @RequestMapping("/productos")
 @CrossOrigin
 public class ProductoController {
-
     @Autowired
     ProductoRepository productoRepository;
 
-    @GetMapping("/listar")
-    public Collection<Producto> listProductos() {return (Collection<Producto>) productoRepository.findAll();}
+    @GetMapping()
+    public ResponseEntity<Collection<Producto>> listaProductos() {return ResponseEntity.ok((Collection<Producto>) productoRepository.findAll());}
 
     @DeleteMapping("/{id}")
-    public String deleteProducto(@PathVariable Long id) {
-
-        productoRepository.deleteById(id);
-
-        return "Se ha eliminado exitosamente el producto";
-
+    public ResponseEntity<String> deleteProducto(@PathVariable Long id) {
+        ResponseEntity<String> response = null;
+        if (productoRepository.findById(id).isPresent()) {
+            productoRepository.deleteById(id);
+            response = ResponseEntity.status(HttpStatus.NO_CONTENT).body("Eliminado");
+        } else {
+            response = ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+        return response;
     }
 
     @GetMapping("/listarporfecha/{fecha1}/{fecha2}")
